requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.3
python_bcrypt==0.3.2
streamlit==1.38.0
~treamlit==1.36.0
